+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `earliest` transformation: finds earliest date among metadata [PR #9](https://github.com/phac-nml/metadatatransformation/pull/9)
 - `populate` transformation: populates an output column with a specific value [PR #10](https://github.com/phac-nml/metadatatransformation/pull/10)
 
+### `Fixed`
+
+- A crash when calculating age if an entire date column contained only floats.
+
 ## [1.0.0] - 2025/03/07
 
 Initial release of phac-nml/metadatatransformation.