+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0]
+## [1.1.0] - 2025/03/17
 
 ### `Added`
 
@@ -30,3 +30,4 @@ Initial release of phac-nml/metadatatransformation.
 ### `Deprecated`
 
 [1.0.0]: https://github.com/phac-nml/metadatatransformation/releases/tag/1.0.0
+[1.1.0]: https://github.com/phac-nml/metadatatransformation/releases/tag/1.1.0